requirements.txt
Flask
flask-cors
requests
gunicorn
psycopg2-binary
python-dotenv
